Stimulus controllers refactoring
