+++ app/views/shared/_profile_buttons.html.erb
@@ -7,18 +7,18 @@
     </svg>
   </button>
 
-  <div class="relative ml-3" data-controller="profile-button">
+  <div class="relative ml-3" data-controller="transition">
     <div>
       <button type="button" class="flex rounded-full bg-white text-sm focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2"
               id="user-menu-button" aria-expanded="false" aria-haspopup="true"
-              data-action="profile-button#toggleProfileMenu">
+              data-action="transition#toggleProfileMenu">
         <span class="sr-only">Open user menu</span>
         <%= image_tag(user.avatar_url, class: "h-8 w-8 rounded-full") %>
       </button>
     </div>
     <div class="hidden absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none"
          role="menu" aria-orientation="vertical" aria-labelledby="user-menu-button" tabindex="-1"
-          data-profile-button-target="profileMenu"
+          data-transition-target="transitionable"
           data-transition-enter="transition ease-out duration-200"
           data-transition-enter-start="transform opacity-0 scale-95"
           data-transition-enter-end="transform opacity-100 scale-100"
